backend/requirements.txt
fastapi[all]>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.21
sqlalchemy>=2.0.0
python-multipart>=0.0.9
openpyxl>=3.1.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
PyJWT>=2.8.0
bcrypt>=4.0.0
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.27
